# Allow multiple pings over one connection