Tiny one. This is so that pinging a rust-libp2p node from ipfs works even without passing -n 1

Before:
```
$ ipfs ping /ip4/127.0.0.1/tcp/33542/ipfs/12D3KooWRpXY8AwbhF9syneVPb6cyrsC9KYSjRywLPWb4mRNaqMu
PING 12D3KooWRpXY8AwbhF9syneVPb6cyrsC9KYSjRywLPWb4mRNaqMu.
Pong received: time=6.79 ms
Ping error: EOF
Ping error: stream reset
Ping error: stream reset
```

After:
```
$ ipfs ping  /ip4/127.0.0.1/tcp/39053/ipfs/12D3KooWFewRGcMQTDJEtYrCkAFfqXDEBQNG7BHgisfJHS9Fqt2S
PING 12D3KooWFewRGcMQTDJEtYrCkAFfqXDEBQNG7BHgisfJHS9Fqt2S.
Pong received: time=2.01 ms
Pong received: time=1.35 ms
Pong received: time=2.62 ms
Pong received: time=2.00 ms
Pong received: time=2.33 ms
Pong received: time=2.10 ms
Pong received: time=2.16 ms
Pong received: time=2.13 ms
Pong received: time=2.43 ms
Pong received: time=1.52 ms
Average latency: 2.07ms
```

Could not find a spec, so I went with this: https://github.com/libp2p/go-libp2p/blob/master/p2p/protocol/ping/ping.go