@tomaka 2020-02-03
I don't think we need to put any timeout, as this is only about receiving pings.