@rklaehn 2020-02-03
I just added the timeout for being identical to the go-ipfs version, which is the closest thing we have to a spec. But I will remove it.